refactor: move usage provider schema to shared
